==> 0X_Effort_LPO.Rmd <==
---
title: "Effort jeu de données faune-france"
author: "Jean N"
date: "24/02/2020"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Objectif du document : Acquisition de données d'effort d'échantillonnage des participants à la base de données faune-france, selon la période mars 2017/ décembre 2019.

### Filtrage des données
  + Uniquement des données de listes complètes
  + Retrait des protocoles SHOC / STOC EPS / STOC MONTAGNE / STOC_ONF / STOC_SITES / WATERBIRDS

```{r import_prep_data,include=FALSE,echo=FALSE}
# chemin
  setwd("C:/git/epoc/data")

  library(lubridate)
  library(ggplot2)
  library(dplyr)


epoc <- read.table(file = paste0(sub("/data","/output",getwd()),"/export_2017_2019.txt"),header=T,sep="\t", dec=","
                   , encoding="UTF-8",quote="")

```

#### Sélection des listes complètes / Retrait des protocoles

```{r filtre, include=TRUE,echo=TRUE}
# filtrage selon listes completes :
      epoc.filt1 <- epoc[epoc$Liste_complete == 1,]
      rm(epoc)

    # filtrage selon les protocoles
      search.prot <- grep("SHOC|STOC_EPS|STOC_MONTAGNE|STOC_ONF|STOC_SITES|WATERBIRDS",epoc.filt1$Protocole) # Selection des lignes contenant les protocoles a retirer
      epoc.filt2 <- epoc.filt1[-search.prot,] # formation d'un dtf sans les lignes selectionnees
```

### Calcul du temps d'écoute pour chaque observation
Pour calculer le temps d'écoute, j'ai converti les données horaires en minutes (ex : 8h20 --> 500 minutes) sur les deux colonnes Heure début et Heure fin. 
Le temps d'écoute correspond à : Heure fin - Heure début
J'ai ensuite sélectionné la première ligne de chaque liste afin d'éviter les doublons.
```{r filt, include=TRUE,echo=TRUE}
# function to convert time observation to minutes since midnight
    time_to_decimal <- function(x) { # need le package lubridate
      x <- hm(x, quiet = TRUE)
      hour(x)*60 + minute(x)
    }

epoc3 <- epoc.filt2
  # formation d'une colonne temps d'ecoute
    # passer le temps comme une variable continue (Heure:minute --> minutes)
      epoc3$Heure_debut <- time_to_decimal(epoc3$Heure_debut) 
      epoc3$Heure_fin <- time_to_decimal(epoc3$Heure_fin)
      
    # tps ecoute = tps fin - tps debut
      epoc3[,"Tps_ecoute"] <- abs(epoc3[,"Heure_fin"] - epoc3[,"Heure_debut"]) # valeur absolue 
      
  # selection des listes (prend uniquement 1 observation par liste == retrait des doublons)
      
      id.liste <- duplicated(epoc3$ID_liste)
      id.liste <- which(id.liste == FALSE)
      
      epoc.liste <- epoc3[id.liste,]
      
```

### Résultats

```{r results_tps, include=TRUE,echo=FALSE}
# Calcul
    # Tps d'ecoute moyen
      tps.mean <- mean(epoc.liste$Tps_ecoute) # 30.18826
      tps.sd <- sd(epoc.liste$Tps_ecoute)
      tps.var <- var(epoc.liste$Tps_ecoute)
      
    # Tps d'ecoute total
      tps.tot <- sum(epoc.liste$Tps_ecoute)
      tps.tot.heur <- tps.tot/60
      tps.tot.jour <- (tps.tot/60)/24
      tps.tot.week <- ((tps.tot/60)/24)/7
      tps.tot.ans <- (((tps.tot/60)/24)/7)/52
      
# Repartition selon ans 2017/2018/2019 -----
  # 2017 ----
    dtf_2017 <- epoc.liste[epoc.liste$Annee == 2017,]
    obs_2017 <- epoc3[epoc3$Annee == 2017,]
      # Nb listes
        liste_2017 <- nrow(dtf_2017)
      # Tps d'ecoute
        # Tps d'ecoute moyen
            tps.mean.2017 <- mean(dtf_2017$Tps_ecoute) # 30.18826
            tps.sd.2017 <- sd(dtf_2017$Tps_ecoute)
            tps.var.2017 <- var(dtf_2017$Tps_ecoute)
            
          # Tps d'ecoute total
            tps.tot.2017 <- sum(dtf_2017$Tps_ecoute)
            tps.tot.heur.2017 <- tps.tot.2017/60
            tps.tot.jour.2017 <- (tps.tot.2017/60)/24
            tps.tot.week.2017 <- ((tps.tot.2017/60)/24)/7
            tps.tot.ans.2017 <- (((tps.tot.2017/60)/24)/7)/52 
            
    # 2018 ----
      dtf_2018 <- epoc.liste[epoc.liste$Annee == 2018,]
      obs_2018 <- epoc3[epoc3$Annee == 2018,]
        # Nb listes
          liste_2018 <- nrow(dtf_2018)
        # Tps d'ecoute
          # Tps d'ecoute moyen
            tps.mean.2018 <- mean(dtf_2018$Tps_ecoute) # 30.18826
            tps.sd.2018 <- sd(dtf_2018$Tps_ecoute)
            tps.var.2018 <- var(dtf_2018$Tps_ecoute)
            
          # Tps d'ecoute total
            tps.tot.2018 <- sum(dtf_2018$Tps_ecoute)
            tps.tot.heur.2018 <- tps.tot.2018/60
            tps.tot.jour.2018 <- (tps.tot.2018/60)/24
            tps.tot.week.2018 <- ((tps.tot.2018/60)/24)/7
            tps.tot.ans.2018 <- (((tps.tot.2018/60)/24)/7)/52 
      
      # 2019 ----
        dtf_2019 <- epoc.liste[epoc.liste$Annee == 2019,]
        obs_2019 <- epoc3[epoc3$Annee == 2019,]
          # Nb listes
            liste_2019 <- nrow(dtf_2019)
          # Tps d'ecoute
            # Tps d'ecoute moyen
            tps.mean.2019 <- mean(dtf_2019$Tps_ecoute) # 30.18826
            tps.sd.2019 <- sd(dtf_2019$Tps_ecoute)
            tps.var.2019 <- var(dtf_2019$Tps_ecoute)
            
            # Tps d'ecoute total
            tps.tot.2019 <- sum(dtf_2019$Tps_ecoute)
            tps.tot.heur.2019 <- tps.tot.2019/60
            tps.tot.jour.2019 <- (tps.tot.2019/60)/24
            tps.tot.week.2019 <- ((tps.tot.2019/60)/24)/7
            tps.tot.ans.2019 <- (((tps.tot.2019/60)/24)/7)/52    
        
      cat("Informations sur les données traités :\n -\tNombre d'observations :",nrow(epoc.filt2),
          "\n -\tNombre de listes :",nrow(epoc.liste))
      
      cat("L'effort d'échantillonnage de tous les protocoles de la base de données faune-france (en dehors des STOC, SHOC et WATERBIRDS) est de :\n",
          round(tps.tot,2)," minutes\n soit :",round(tps.tot.heur,2)," heures\n soit :",round(tps.tot.jour,2)," jours\n soit :",
          round(tps.tot.week,2)," semaines\n soit :",round(tps.tot.ans,2),"ans")
      cat("Temps moyen d'écoute par liste complète :",round(tps.mean,2))
      
            cat("Pour l'année 2017, nous avons :\n -\t",nrow(obs_2017)," Observations\n -\t",
          liste_2017," Listes réalisées\n -\t",round(tps.mean.2017)," minutes de temps d'écoute moyen\n -\t",
          tps.tot.2017," minutes de temps d'écoute total soit :\n \t-\t",round(tps.tot.heur.2017,2)," Heures\n \t-\t",
          tps.tot.jour.2017," Jours\n \t-\t",tps.tot.week.2017," Semaines\n \t-\t",tps.tot.ans.2017," Années")
      
      cat("Pour l'année 2018, nous avons :\n -\t",nrow(obs_2018)," Observations\n -\t",
          liste_2018," Listes réalisées\n -\t",round(tps.mean.2018)," minutes de temps d'écoute moyen\n -\t",
          tps.tot.2018," minutes de temps d'écoute total soit :\n \t-\t",round(tps.tot.heur.2018,2)," Heures\n \t-\t",
          tps.tot.jour.2018," Jours\n \t-\t",tps.tot.week.2018," Semaines\n \t-\t",tps.tot.ans.2018," Années")
      
      cat("Pour l'année 2019, nous avons :\n -\t",nrow(obs_2019)," Observations\n -\t",
          liste_2019," Listes réalisées\n -\t",round(tps.mean.2019)," minutes de temps d'écoute moyen\n -\t",
          tps.tot.2019," minutes de temps d'écoute total soit :\n \t-\t",round(tps.tot.heur.2019,2)," Heures\n \t-\t",
          tps.tot.jour.2019," Jours\n \t-\t",tps.tot.week.2019," Semaines\n \t-\t",tps.tot.ans.2019," Années")
```

```{r results_tps2, include=TRUE, echo=FALSE, warning=FALSE}
rm(epoc.filt1)
rm(epoc.filt2)

  # boxplot des tps d'ecoute selon les mois/annees
      epoc3 %>%
        mutate(Mois2 = as.Date(paste0("2015-", Mois,"-01"),"%Y-%m-%d")) %>%
        ggplot(aes(x = Mois2, y = Tps_ecoute)) +
        geom_boxplot(aes(group=Mois), fill = "darkolivegreen3",outlier.shape = NA) +
        scale_y_continuous(limits = quantile(epoc.liste$Tps_ecoute, c(0.1,0.90)))  +
        facet_wrap(~ Annee, ncol = 3) +
        labs(title = "Repartition annuelle du temps d'écoute d'observations",
             subtitle = "Listes completes faune-france",
             y = "Temps d'ecoute (minutes)",
             x = "Mois") + theme_bw(base_size = 15) +
        scale_x_date(date_labels = "%b")
```
